+++ bin/light-base/src/json_rpc_service/state_chain.rs
@@ -23,6 +23,7 @@ use crate::runtime_service;
 
 use futures::{lock::MutexGuard, prelude::*};
 use smoldot::{
+    executor::runtime_host,
     header,
     json_rpc::{self, methods, requests_subscriptions},
     network::protocol,
@@ -832,6 +833,49 @@ impl<TPlat: Platform> Background<TPlat> {
 
         self.requests_subscriptions
             .respond(&state_machine_request_id, response)
+            .await;
+    }
+
+    /// Handles a call to [`methods::MethodCall::state_call`].
+    pub(super) async fn state_call(
+        self: &Arc<Self>,
+        request_id: &str,
+        state_machine_request_id: &requests_subscriptions::RequestId,
+        function_to_call: &str,
+        call_parameters: methods::HexString,
+        hash: Option<methods::HashHexString>,
+    ) {
+        let block_hash = if let Some(hash) = hash {
+            hash.0
+        } else {
+            header::hash_from_scale_encoded_header(
+                &sub_utils::subscribe_best(&self.runtime_service).await.0,
+            )
+        };
+
+        let result = self
+            .runtime_call(
+                &block_hash,
+                function_to_call,
+                iter::once(call_parameters.0),
+                3,
+                Duration::from_secs(10),
+                NonZeroU32::new(3).unwrap(),
+            )
+            .await;
+
+        let response = match result {
+            Ok(data) => methods::Response::state_call(methods::HexString(data.to_vec()))
+                .to_json_response(request_id),
+            Err(error) => json_rpc::parse::build_error_response(
+                request_id,
+                json_rpc::parse::ErrorResponse::ServerError(-32000, &error.to_string()),
+                None,
+            ),
+        };
+
+        self.requests_subscriptions
+            .respond(state_machine_request_id, response)
             .await;
     }